created hash utils
